create ClientOptionsTrait
